scripts/requirements.txt
requests
feedparser
beautifulsoup4
python-dateutil